requirements.txt
# Base dependencies
wheel>=0.38.4
setuptools>=65.5.1

# Essential numerical and data processing
numpy>=1.24.0
pandas>=2.0.0
scipy==1.11.3

# Machine Learning Core
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm==4.1.0
catboost==1.2.2
optuna==3.3.0
shap==0.42.1

# Optimization and Benchmarking
tabulate>=0.9.0
protobuf==3.20.0  # For compatibility

# Dependencies for GPU support (optional)
# cupy>=12.0.0  # Uncomment if using NVIDIA GPU
# cuda-python>=12.0  # Uncomment if using NVIDIA GPU

# Visualization
plotly>=5.18.0
seaborn==0.12.2
matplotlib==3.7.3
wordcloud>=1.9.0
kaleido==0.2.1

# Web App
streamlit>=1.28.0
streamlit-lottie==0.0.5

# Data Analysis & Profiling
ydata-profiling>=4.6.0
visions==0.7.4
tangled-up-in-unicode==0.2.0
phik==0.12.3
tqdm>=4.48.2
htmlmin>=0.1.12
imagehash>=4.0
multimethod>=1.4

# MLOps
mlflow==2.5.0

# Jupyter Environment
jupyter==1.0.0
ipykernel>=6.22.0
nbformat>=5.9.0
nbconvert>=7.9.0
notebook>=6.4.12
jupytext==1.15.0

# Utilities
pillow==10.0.1
requests==2.31.0
joblib>=1.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2>=2.11.1
networkx>=2.5
